Change route api and Create StoreMessageRequest.php and fix some errors

+++ app/Http/Controllers/CategoriesController.php
@@ -6,7 +6,6 @@ use App\Http\Requests\Category\StoreCategoryRequest;
 use App\Http\Requests\Category\UpdateCategoryRequest;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
-use Exception;
 use Spatie\FlareClient\Http\Exceptions\BadResponseCode;
 
 class CategoriesController extends Controller
@@ -16,9 +15,12 @@ class CategoriesController extends Controller
      */
     public function index()
     {
+        // Get all categories paginated
+        $categories = Category::paginate();
 
-        return CategoryResource::collection(Category::paginate());
-
+        return $categories->count() == 1
+            ? new CategoryResource($categories->first()) // return a single category
+            : CategoryResource::collection($categories); // return a collection of categories
     }
 
     /**
@@ -26,36 +28,59 @@ class CategoriesController extends Controller
      */
     public function store(StoreCategoryRequest $request)
     {
+        // Validate the request data
         $valid_data = $request->validated();
+
+        // Create the category
         $category = Category::create($valid_data);
 
+        // Return the newly created category
         return new CategoryResource($category);
-
-
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Category $category)
+    public function show(string $id)
     {
-        return new CategoryResource($category);
+        // Get the specified category
+        $category = Category::find($id);
 
+        // Check if the category exists
+        if (!$category) {
+            return response()->json(['message' => 'Category ID not found'], 404);
+        }
+
+        // Return the specified category
+        return new CategoryResource($category);
     }
 
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateCategoryRequest $request, Category $category)
+    public function update(UpdateCategoryRequest $request, string $id)
     {
+
+        // Get the category
+        $category = Category::find($id);
+
+        // Check if the category exists
+        if (!$category) {
+            return response()->json(['message' => 'Category ID not found'], 404);
+        }
+
         try {
+            // Validate the request
             $valid_data = $request->validated();
+
+            // Update the category
             $category->update($valid_data);
 
+            // Return the category
             return new CategoryResource($category);
-
         } catch (BadResponseCode $e) {
+            // Return the error message if category id equals parent id
             return response()->json(['error' => $e->getMessage()], 422);
         }
 
@@ -64,9 +89,20 @@ class CategoriesController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Category $category)
+    public function destroy(string $id)
     {
+        // Get the category
+        $category = Category::find($id);
+
+        // Check if the category exists
+        if (!$category) {
+            return response()->json(['message' => 'Category ID not found'], 404);
+        }
+
+        // Delete the category
         $category->delete();
-        return response()->noContent();
+
+        // Return message success
+        return response()->json(['message' => 'Deleted successfully']);
     }
 }
